backend/requirements.txt
flask
flask-cors
requests
pandas
numpy
python-dotenv
supabase
sentence_transformers
cassandra-driver
kafka-python

